requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5,<3
pydantic-settings>=2,<3
python-multipart==0.0.6


# LLM & RAG
sentence-transformers==2.3.1
faiss-cpu==1.8.0
requests==2.31.0

# Tools
duckduckgo_search==4.1.1
radon==6.0.1

# Frontend
streamlit==1.30.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
loguru==0.7.2

# Utils
python-dotenv==1.0.0

pydantic>=2.5.0
